+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -144,6 +144,9 @@ public enum FeatureFlag: String, CaseIterable {
     /// Enable Disable the use of suggested folders
     case suggestedFolders
 
+    /// Parse toc value in chapters
+    case parseChaptersToc
+
     /// Enable the generated transcript
     case generatedTranscripts
 
@@ -246,6 +249,8 @@ public enum FeatureFlag: String, CaseIterable {
             true
         case .suggestedFolders:
             false
+        case .parseChaptersToc:
+            true
         case .generatedTranscripts:
             false
         case .podcastViewChanges: